move ellipsis detector hook to its own folder
