+++ src/app/custom-hooks/ellipsis-detector/index.ts
