+++ src/app/custom-hooks/ellipsis-detector/useEllipsisDetector.ts
